New workflow for update service
Using the new promise architecture, implement the workflow for updating the data from the moves api.

Right now this implementation takes a bit of paper to work out the workflow, so in the future a new implementation that wraps several of the design concepts should be developed.

The only missing part of the implementation is the lookup and update/create of the nodes for the data in the API.